refactor: move all shareLink files to components, inluding assets
