--- apps/sports-helsinki/src/common-events/components/shareLinks/shareLinks.module.scss
@@ -1,57 +1,0 @@
-.shareSubSection {
-  padding: var(--spacing-xl) 0;
-}
-
-.shareSubSectionTitle {
-  font-size: var(--fontsize-body-l);
-  font-weight: 600;
-  margin: 1rem 0;
-}
-
-.shareLinkList {
-  list-style: none;
-  margin: unset;
-  padding: unset;
-  display: flex;
-}
-
-.shareLinkItem {
-  margin-right: var(--spacing-m);
-
-  &:last-child {
-    margin-right: unset;
-  }
-
-  button {
-    line-height: 1;
-
-    &:focus {
-      outline: 2px solid var(--color-primary-black);
-    }
-  }
-}
-
-.relativePosition {
-  position: relative;
-}
-
-.copyButton {
-  fill: var(--color-primary-black);
-}
-
-.linkCopyButtonSuccess {
-  fill: var(--color-success);
-}
-
-.successTooltip {
-  position: absolute;
-  top: calc(100% + 0.5rem);
-  left: 0;
-  padding: 0.25rem 0.6rem;
-
-  font-size: var(--fontsize-body-s);
-  color: var(--color-white);
-  white-space: nowrap;
-
-  background-color: var(--color-success);
-}
