+++ packages/components/src/components/shareLinks/LinkedInShareLink.tsx
@@ -1,9 +1,9 @@
-import type { ShareLinkProps } from 'events-helsinki-components';
-import { ShareLinkBase } from 'events-helsinki-components';
 import { useTranslation } from 'next-i18next';
 import React from 'react';
+import type { ShareLinkProps } from 'events-helsinki-components';
+import { ShareLinkBase } from 'events-helsinki-components';
 
-import LinkedInIcon from '../../../assets/icons/svg/linkedin.svg';
+import LinkedInIcon from '../../assets/icons/svg/linkedin.svg';
 
 const linkedInShareUrl = 'https://linkedin.com/shareArticle';
