+++ packages/components/src/components/shareLinks/ShareLinks.tsx
@@ -1,12 +1,12 @@
 import classNames from 'classnames';
+import React from 'react';
 import {
   isClient,
   CopyButton,
   useCommonTranslation,
 } from 'events-helsinki-components';
-import React from 'react';
 
-import LinkIcon from '../../../assets/icons/svg/link.svg';
+import LinkIcon from '../../assets/icons/svg/link.svg';
 import FacebookShareLink from './FacebookShareLink';
 import LinkedInShareLink from './LinkedInShareLink';
 import styles from './shareLinks.module.scss';
@@ -18,7 +18,6 @@ export interface ShareLinksProps {
 
 const ShareLinks: React.FC<ShareLinksProps> = ({ title }) => {
   const { t } = useCommonTranslation();
-
   // We are using the client only accessible href. By doing this, we do not need
   // to pass the original request from the server. This same pattern was used in
   // MyHelsinki. Limitation is that sharing buttons will be re-rendered on client
