+++ packages/components/src/components/shareLinks/__tests__/FacebookShareLink.test.tsx
@@ -1,8 +1,6 @@
-import type { ShareLinkProps } from 'events-helsinki-components/components';
-import React from 'react';
 import { render, screen } from '@/test-utils';
-
-import FacebookShareLink from '../FacebookShareLink';
+import type { ShareLinkProps } from 'events-helsinki-components';
+import FacebookShareLink from 'events-helsinki-components/src/components/shareLinks/FacebookShareLink';
 
 const renderComponent = (props: ShareLinkProps) =>
   render(<FacebookShareLink {...props} />);
