+++ packages/components/src/components/shareLinks/__tests__/LinkedInShareLink.test.tsx
@@ -1,6 +1,5 @@
-import type { ShareLinkProps } from 'events-helsinki-components/components';
-import React from 'react';
 import { render, screen } from '@/test-utils';
+import type { ShareLinkProps } from 'events-helsinki-components';
 import LinkedInShareLink from '../LinkedInShareLink';
 
 const renderComponent = (props: ShareLinkProps) =>
