+++ packages/components/src/components/shareLinks/__tests__/ShareLinks.test.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
-
 import { render, screen } from '@/test-utils';
+
 import type { ShareLinksProps } from '../ShareLinks';
 import ShareLinks from '../ShareLinks';
