+++ packages/components/src/components/shareLinks/__tests__/TwitterShareLink.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render, screen } from '@/test-utils';
 import TwitterShareLink from '../TwitterShareLink';
